// apps/editor/src/core/layout/types.ts
export type LayoutCommandType =
  | "mark"
  | "markStart"
  | "markMiddle"
  | "markEnd"
  | "markLineStart"
  | "markLineMiddle"
  | "markLineEnd"
  | "markChar"
  | "goto"
  | "resume"
  | "offset"
  | "offsetX"
  | "offsetY"
  | "set"
  | "group"
  | "right"
  | "left"
  | "down"
  | "up"
  | "lineStart"
  | "lineEnd"
  | "moveChar"
  | "pushDisplayOffset"
  | "popDisplayOffset"
  | "flow";
// ...
export interface LayoutCommand {
  isCommand: true;
  type: LayoutCommandType;
  params: any;
}